🚚 MainStyle 코드 파일 분리하여 별도 관리
